AWS: improve compatibility of lab_id for resource naming
This commit controls the characterset used for the autogenerated lab_id,
which avoids problems when the lab_id is used as part of the 'name'
attribute on an AWS resource with a narrower range of supported
characters.

While the random_id.b64_url property enforces a URL-safe characterset,
it also includes underscores, which are not allowed for S3 buckets.

Despite the terraform recommendation to use random_id over random_string,
the random_string resource will continue to exist for backward compatibility
and it's configurable options provide the guarantees needed for not just
URL-safe strings but also AWS-resource-name-safe strings.